Rename components and fix add selected from typeahead, and clean selection after close and generation
