+++ admin-dev/themes/new-theme/js/pages/product/edit/combinations-manager.js
@@ -32,7 +32,7 @@ import ProductEventMap from '@pages/product/product-event-map';
 import initCombinationModal from '@pages/product/components/combination-modal';
 import initFilters from '@pages/product/components/filters';
 import ConfirmModal from '@components/modal';
-import initCombinationGenerator from '@pages/product/components/combinations';
+import initCombinationGenerator from '@pages/product/components/generator';
 
 const {$} = window;
 const CombinationEvents = ProductEventMap.combinations;
